refactor (gravsearch): Continue refactoring Gravsearch classes.
